Implementacion de modal para mostrar usuario

+++ back_front/src/app/components/cards-products/cards-products.component.ts
@@ -8,31 +8,54 @@ import { ApiService } from 'src/app/services/api.service';
 })
 export class CardsProductsComponent  implements OnInit {
   public resp: productos[] = [];
-
+  isLoading = false; // Controla el estado de carga
+  error: string | null = null; // Almacena un mensaje de error si ocurre
   constructor( private newService: ApiService) { }
 
   ngOnInit() {
-    this.newService.getTopHeadlines()
-      .subscribe(resp => {
-        console.log(resp);
+    this.fetchProducts();
+  }
+
+  fetchProducts() {
+    this.isLoading = true; // Comienza la carga
+    this.newService.getTopHeadlines().subscribe({
+      next: (resp) => {
+        this.isLoading = false; // Detiene la carga cuando los datos son recibidos
         if (Array.isArray(resp)) {
-          this.resp = resp; 
+          this.resp = resp;
         } else {
-          this.resp = [resp]; 
+          this.resp = [resp];
         }
-      });
+      }, error: (error) => {
+        this.isLoading = false; // Detiene la carga si ocurre un error
+        this.error = "No se pudieron cargar los productos."; // Mensaje de error
+        console.error(error);
+      }
+    });
   }
 
   eliminarDato(id_producto: number) {
-    this.newService.eliminarDato(id_producto).subscribe(
-      (response: any) => {
+    const confirmacion = confirm("¿Estás seguro de que deseas eliminar este producto?");
+    if (!confirmacion) {
+      return;
+    }
+    
+    this.newService.eliminarDato(id_producto).subscribe({
+      next: (response) => {
         console.log(response);
         this.resp = this.resp.filter(item => item.id_producto !== id_producto);
       },
-      error => {
+      error: (error) => {
         console.error("Error al eliminar el registro:", error);
+        this.error = "Error al eliminar el producto."; // Asume que tienes una manera de mostrar esto en tu UI
       }
-    );
+    });
   }
-
+  
+  confirmarEliminacion(id_producto: number) {
+    const confirmacion = confirm("¿Estás seguro de que deseas eliminar este producto?");
+    if (confirmacion && id_producto !== undefined) {
+      this.eliminarDato(id_producto);
+    }
+  }
 }
